requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
altair==5.3.0
    # via streamlit
attrs==23.2.0
    # via
    #   jsonschema
    #   referencing
black==24.4.2
    # via -r requirements.in
blinker==1.8.2
    # via streamlit
build==1.2.1
    # via pip-tools
cachetools==5.3.3
    # via streamlit
certifi==2024.6.2
    # via requests
cfgv==3.4.0
    # via pre-commit
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   black
    #   pip-tools
    #   streamlit
colorama==0.4.6
    # via
    #   build
    #   click
    #   tqdm
contourpy==1.2.1
    # via matplotlib
cycler==0.12.1
    # via matplotlib
distlib==0.3.8
    # via virtualenv
filelock==3.15.4
    # via virtualenv
fonttools==4.53.0
    # via matplotlib
gitdb==4.0.11
    # via gitpython
gitpython==3.1.43
    # via streamlit
identify==2.5.36
    # via pre-commit
idna==3.7
    # via requests
jinja2==3.1.4
    # via
    #   altair
    #   pydeck
joblib==1.4.2
    # via
    #   pynndescent
    #   scikit-learn
jsonschema==4.22.0
    # via altair
jsonschema-specifications==2023.12.1
    # via jsonschema
kiwisolver==1.4.5
    # via matplotlib
llvmlite==0.43.0
    # via
    #   numba
    #   pynndescent
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.5
    # via jinja2
matplotlib==3.9.0
    # via -r requirements.in
mdurl==0.1.2
    # via markdown-it-py
mypy==1.10.0
    # via -r requirements.in
mypy-extensions==1.0.0
    # via
    #   black
    #   mypy
nodeenv==1.9.1
    # via pre-commit
numba==0.60.0
    # via
    #   pynndescent
    #   umap-learn
numpy==2.0.0
    # via
    #   -r requirements.in
    #   altair
    #   contourpy
    #   matplotlib
    #   numba
    #   pandas
    #   pyarrow
    #   pydeck
    #   scikit-learn
    #   scipy
    #   streamlit
    #   umap-learn
packaging==24.1
    # via
    #   altair
    #   black
    #   build
    #   matplotlib
    #   plotly
    #   streamlit
pandas==2.2.2
    # via
    #   -r requirements.in
    #   altair
    #   streamlit
pathspec==0.12.1
    # via black
pillow==10.3.0
    # via
    #   matplotlib
    #   streamlit
pip-tools==7.4.1
    # via -r requirements.in
platformdirs==4.2.2
    # via
    #   black
    #   virtualenv
plotly==5.22.0
    # via -r requirements.in
pre-commit==3.7.1
    # via -r requirements.in
protobuf==5.27.1
    # via streamlit
pyarrow==16.1.0
    # via streamlit
pydeck==0.9.1
    # via streamlit
pygments==2.18.0
    # via rich
pynndescent==0.5.13
    # via umap-learn
pyparsing==3.1.2
    # via matplotlib
pyproject-hooks==1.1.0
    # via
    #   build
    #   pip-tools
python-dateutil==2.9.0.post0
    # via
    #   matplotlib
    #   pandas
pytz==2024.1
    # via pandas
pyyaml==6.0.1
    # via pre-commit
referencing==0.35.1
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.3
    # via
    #   -r requirements.in
    #   streamlit
rich==13.7.1
    # via streamlit
rpds-py==0.18.1
    # via
    #   jsonschema
    #   referencing
ruff==0.4.10
    # via -r requirements.in
scikit-learn==1.5.0
    # via
    #   -r requirements.in
    #   pynndescent
    #   umap-learn
scipy==1.13.1
    # via
    #   -r requirements.in
    #   pynndescent
    #   scikit-learn
    #   umap-learn
six==1.16.0
    # via python-dateutil
smmap==5.0.1
    # via gitdb
streamlit==1.36.0
    # via -r requirements.in
tenacity==8.4.1
    # via
    #   plotly
    #   streamlit
threadpoolctl==3.5.0
    # via scikit-learn
toml==0.10.2
    # via streamlit
toolz==0.12.1
    # via altair
tornado==6.4.1
    # via streamlit
tqdm==4.66.4
    # via umap-learn
typing-extensions==4.12.2
    # via
    #   mypy
    #   streamlit
tzdata==2024.1
    # via pandas
umap-learn==0.5.6
    # via -r requirements.in
urllib3==2.2.2
    # via requests
virtualenv==20.26.3
    # via pre-commit
watchdog==4.0.1
    # via streamlit
wheel==0.43.0
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
